+++ src/Html/HtmlProvider.fs
@@ -27,8 +27,7 @@ type public HtmlProvider(cfg:TypeProviderConfig) as this =
         let sample = args.[0] :?> string
         let preferOptionals = args.[1] :?> bool
         let includeLayoutTables = args.[2] :?> bool
-        let missingValues = args.[3] :?> string
-        let missingValuesList = missingValues.Split([| ',' |], StringSplitOptions.RemoveEmptyEntries)
+        let missingValuesStr = args.[3] :?> string
         let cultureStr = args.[4] :?> string
         let cultureInfo = TextRuntime.GetCulture cultureStr
         let resolutionFolder = args.[5] :?> string
@@ -40,8 +39,8 @@ type public HtmlProvider(cfg:TypeProviderConfig) as this =
             let htmlType = using (IO.logTime "Inference" sample) <| fun _ ->
                 tables 
                 |> List.filter (fun table -> table.Headers.Length > 0)
-                |> List.map (fun table -> table.Name, HtmlInference.inferColumns preferOptionals missingValuesList cultureInfo table.Headers table.Rows)
-                |> HtmlGenerator.generateTypes asm ns typeName (missingValues, cultureStr) replacer
+                |> List.map (fun table -> table.Name, HtmlInference.inferColumns preferOptionals (TextRuntime.GetMissingValues missingValuesStr) cultureInfo table.Headers table.Rows)
+                |> HtmlGenerator.generateTypes asm ns typeName (missingValuesStr, cultureStr) replacer
 
             using (IO.logTime "TypeGeneration" sample) <| fun _ ->
 
@@ -53,25 +52,23 @@ type public HtmlProvider(cfg:TypeProviderConfig) as this =
         generateType "HTML" sample (*sampleIsList*)false (fun _ -> HtmlDocument.Parse >> HtmlRuntime.getTables includeLayoutTables) (fun _ _ -> failwith "Not Applicable")
                      getSpecFromSamples version this cfg replacer resolutionFolder typeName
 
-    let defaultMissingValues = String.Join(",", TextConversions.DefaultMissingValues)
-
     // Add static parameter that specifies the API we want to get (compile-time) 
     let parameters = 
         [ ProvidedStaticParameter("Sample", typeof<string>, parameterDefaultValue = "")           
           ProvidedStaticParameter("PreferOptionals", typeof<bool>, parameterDefaultValue = false)
           ProvidedStaticParameter("IncludeLayoutTables", typeof<bool>, parameterDefaultValue = false)
-          ProvidedStaticParameter("MissingValues", typeof<string>, parameterDefaultValue = defaultMissingValues)
+          ProvidedStaticParameter("MissingValues", typeof<string>, parameterDefaultValue = "")
           ProvidedStaticParameter("Culture", typeof<string>, parameterDefaultValue = "")
           ProvidedStaticParameter("ResolutionFolder", typeof<string>, parameterDefaultValue = "") ]
   
     let helpText = 
-        """<summary>Typed representation of an HTML file</summary>
-           <param name='Sample'>Location of an HTML sample file or a string containing a sample HTML document</param>
-           <param name='PreferOptionals'>When set to true, inference will prefer to use the option type instead of nullable types, double.NaN or "" for missing values. Defaults to false</param>
+        """<summary>Typed representation of an HTML file.</summary>
+           <param name='Sample'>Location of an HTML sample file or a string containing a sample HTML document.</param>
+           <param name='PreferOptionals'>When set to true, inference will prefer to use the option type instead of nullable types, `double.NaN` or `""` for missing values. Defaults to false.</param>
            <param name='IncludeLayoutTables'>Includes tables that are potentially layout tables (with cellpadding=0 and cellspacing=0 attributes)</param>
-           <param name='MissingValues'>The set of strings recogized as missing values. Defaults to """ + "\"" + defaultMissingValues + "\"" + """</param>
-           <param name='Culture'>The culture used for parsing numbers and dates. Defaults to the invariant culture</param>
-           <param name='ResolutionFolder'>A directory that is used when resolving relative file references (at design time and in hosted execution)</param>"""
+           <param name='MissingValues'>The set of strings recogized as missing values. Defaults to `""" + String.Join(",", TextConversions.DefaultMissingValues) + """`.</param>
+           <param name='Culture'>The culture used for parsing numbers and dates. Defaults to the invariant culture.</param>
+           <param name='ResolutionFolder'>A directory that is used when resolving relative file references (at design time and in hosted execution).</param>"""
   
     do htmlProvTy.AddXmlDoc helpText
     do htmlProvTy.DefineStaticParameters(parameters, buildTypes)