+++ src/Net/Http.fs
@@ -10,6 +10,7 @@ open System.IO
 open System.Net
 open System.Text
 open System.Reflection
+open System.Runtime.CompilerServices
 open FSharp.Data.Runtime
 
 /// The method to use in an HTTP request
@@ -277,8 +278,53 @@ module private Helpers =
         return output 
     }
 
+    let (?) obj prop =
+#if FX_NET_CORE_REFLECTION
+        let prop = obj.GetType().GetRuntimeProperty(prop)
+        if prop <> null && prop.CanRead then
+            try
+                prop.GetValue(obj) |> unbox |> Some
+            with _ -> 
+                None
+        else
+            None
+#else
+        let prop = obj.GetType().GetProperty(prop)
+        if prop <> null && prop.CanRead then
+            try
+                prop.GetValue(obj, [| |]) |> unbox |> Some
+            with _ -> 
+                None
+        else
+            None
+#endif
+
+    let (?<-) obj prop value =
+#if FX_NET_CORE_REFLECTION
+        let prop = obj.GetType().GetRuntimeProperty(prop)
+        if prop <> null && prop.CanWrite then
+            try 
+                prop.SetValue(obj, box value) |> ignore
+                true
+            with _ -> 
+                false
+        else
+            false
+#else
+        let prop = obj.GetType().GetProperty(prop)
+        if prop <> null && prop.CanWrite then
+            try 
+                prop.SetValue(obj, value, [| |]) |> ignore
+                true
+            with _ -> 
+                false
+        else
+            false
+#endif
+
     let writeBody (req:HttpWebRequest) (postBytes:byte[]) = async { 
 #if FX_NO_WEBREQUEST_CONTENTLENGTH
+        ignore (req?ContentLength <- int64 postBytes.Length)
 #else
         req.ContentLength <- int64 postBytes.Length
 #endif
@@ -337,87 +383,87 @@ module private Helpers =
     let setHeaders headers (req:HttpWebRequest) =
         let hasContentType = ref false
         headers |> Option.iter (checkForRepeatedHeaders [])
-        headers |> Option.iter (List.iter (fun (header, value) ->
-            match header with
-            | "Accept" -> req.Accept <- value
-            | "Accept-Charset" -> req.Headers.[HeaderEnum.AcceptCharset] <- value
-            | "Accept-Datetime" -> req.Headers.["Accept-Datetime"] <- value
-            | "Accept-Encoding" -> req.Headers.[HeaderEnum.AcceptEncoding] <- value
-            | "Accept-Language" -> req.Headers.[HeaderEnum.AcceptLanguage] <- value
-            | "Allow" -> req.Headers.[HeaderEnum.Allow] <- value
-            | "Authorization" -> req.Headers.[HeaderEnum.Authorization] <- value
-            | "Cache-Control" -> req.Headers.[HeaderEnum.CacheControl] <- value
+        headers |> Option.iter (List.iter (fun (header:string, value) ->
+            match header.ToLowerInvariant() with
+            | "accept" -> req.Accept <- value
+            | "accept-charset" -> req.Headers.[HeaderEnum.AcceptCharset] <- value
+            | "accept-datetime" -> req.Headers.["Accept-Datetime"] <- value
+            | "accept-encoding" -> req.Headers.[HeaderEnum.AcceptEncoding] <- value
+            | "accept-language" -> req.Headers.[HeaderEnum.AcceptLanguage] <- value
+            | "allow" -> req.Headers.[HeaderEnum.Allow] <- value
+            | "authorization" -> req.Headers.[HeaderEnum.Authorization] <- value
+            | "cache-control" -> req.Headers.[HeaderEnum.CacheControl] <- value
 #if FX_NO_WEBREQUEST_CONNECTION
-            | "Connection" -> req.Headers.[HeaderEnum.Connection] <- value
+            | "connection" -> if not (req?Connection <- value) then req.Headers.[HeaderEnum.Connection] <- value
 #else
-            | "Connection" -> req.Connection <- value
+            | "connection" -> req.Connection <- value
 #endif
-            | "Content-Encoding" -> req.Headers.[HeaderEnum.ContentEncoding] <- value
-            | "Content-Language" -> req.Headers.[HeaderEnum.ContentLanguage] <- value
-            | "Content-Location" -> req.Headers.[HeaderEnum.ContentLocation] <- value
-            | "Content-MD5" -> req.Headers.[HeaderEnum.ContentMd5] <- value
-            | "Content-Range" -> req.Headers.[HeaderEnum.ContentRange] <- value
-            | "Content-Type" ->
+            | "content-encoding" -> req.Headers.[HeaderEnum.ContentEncoding] <- value
+            | "content-Language" -> req.Headers.[HeaderEnum.ContentLanguage] <- value
+            | "content-Location" -> req.Headers.[HeaderEnum.ContentLocation] <- value
+            | "content-md5" -> req.Headers.[HeaderEnum.ContentMd5] <- value
+            | "content-range" -> req.Headers.[HeaderEnum.ContentRange] <- value
+            | "content-type" ->
                 req.ContentType <- value
                 hasContentType := true
 #if FX_NO_WEBREQUEST_DATE
-            | "Date" -> req.Headers.[HeaderEnum.Date] <- value
+            | "date" -> if not (req?Date <- DateTime.ParseExact(value, "R", CultureInfo.InvariantCulture)) then req.Headers.[HeaderEnum.Date] <- value
 #else
-            | "Date" -> req.Date <- DateTime.ParseExact(value, "R", CultureInfo.InvariantCulture)
+            | "date" -> req.Date <- DateTime.ParseExact(value, "R", CultureInfo.InvariantCulture)
 #endif
 #if FX_NO_WEBREQUEST_EXPECT
-            | "Expect" -> req.Headers.[HeaderEnum.Expect] <- value
+            | "expect" -> if not (req?Expect <- value) then req.Headers.[HeaderEnum.Expect] <- value
 #else
-            | "Expect" -> req.Expect <- value
+            | "expect" -> req.Expect <- value
 #endif
-            | "Expires" -> req.Headers.[HeaderEnum.Expires] <- value
-            | "From" -> req.Headers.[HeaderEnum.From] <- value
+            | "expires" -> req.Headers.[HeaderEnum.Expires] <- value
+            | "from" -> req.Headers.[HeaderEnum.From] <- value
 #if FX_NO_WEBREQUEST_HOST
-            | "Host" -> req.Headers.[HeaderEnum.Host] <- value
+            | "host" -> if not (req?Host <- value) then req.Headers.[HeaderEnum.Host] <- value
 #else
-            | "Host" -> req.Host <- value
+            | "host" -> req.Host <- value
 #endif       
-            | "If-Match" -> req.Headers.[HeaderEnum.IfMatch] <- value
+            | "if-match" -> req.Headers.[HeaderEnum.IfMatch] <- value
 #if FX_NO_WEBREQUEST_IFMODIFIEDSINCE
-            | "IfModifiedSince" -> req.Headers.[HeaderEnum.IfModifiedSince] <- value
+            | "if-modified-since" -> if not (req?IfModifiedSince <- DateTime.ParseExact(value, "R", CultureInfo.InvariantCulture)) then req.Headers.[HeaderEnum.IfModifiedSince] <- value
 #else
-            | "If-Modified-Since" -> req.IfModifiedSince <- DateTime.ParseExact(value, "R", CultureInfo.InvariantCulture)
+            | "if-modified-since" -> req.IfModifiedSince <- DateTime.ParseExact(value, "R", CultureInfo.InvariantCulture)
 #endif
-            | "If-None-Match" -> req.Headers.[HeaderEnum.IfNoneMatch] <- value
-            | "If-Range" -> req.Headers.[HeaderEnum.IfRange] <- value
-            | "If-Unmodified-Since" -> req.Headers.[HeaderEnum.IfUnmodifiedSince] <- value
-            | "Keep-Alive" -> req.Headers.[HeaderEnum.KeepAlive] <- value
-            | "Last-Modified" -> req.Headers.[HeaderEnum.LastModified] <- value
-            | "Max-Forwards" -> req.Headers.[HeaderEnum.MaxForwards] <- value
-            | "Origin" -> req.Headers.["Origin"] <- value
-            | "Pragma" -> req.Headers.[HeaderEnum.Pragma] <- value
+            | "if-none-match" -> req.Headers.[HeaderEnum.IfNoneMatch] <- value
+            | "if-range" -> req.Headers.[HeaderEnum.IfRange] <- value
+            | "if-unmodified-since" -> req.Headers.[HeaderEnum.IfUnmodifiedSince] <- value
+            | "keep-alive" -> req.Headers.[HeaderEnum.KeepAlive] <- value
+            | "last-modified" -> req.Headers.[HeaderEnum.LastModified] <- value
+            | "max-forwards" -> req.Headers.[HeaderEnum.MaxForwards] <- value
+            | "origin" -> req.Headers.["Origin"] <- value
+            | "pragma" -> req.Headers.[HeaderEnum.Pragma] <- value
 #if FX_NO_WEBREQUEST_RANGE
-            | "Range(start, finish)" -> req.Headers.[HeaderEnum.Range] <- value
+            | "range" -> req.Headers.[HeaderEnum.Range] <- value
 #else
-            | "Range" -> 
+            | "range" -> 
                 if not (value.StartsWith("bytes=")) then failwith "Invalid value for the Range header"
                 let bytes = value.Substring("bytes=".Length).Split('-')
                 if bytes.Length <> 2 then failwith "Invalid value for the Range header"
                 req.AddRange(int64 bytes.[0], int64 bytes.[1])
 #endif
-            | "Proxy-Authorization" -> req.Headers.[HeaderEnum.ProxyAuthorization] <- value
+            | "proxy-authorization" -> req.Headers.[HeaderEnum.ProxyAuthorization] <- value
 #if FX_NO_WEBREQUEST_REFERER
-            | "Referer" -> req.Headers.[HeaderEnum.Referer] <- value
+            | "referer" -> if not (req?Referer <- value) then try req.Headers.[HeaderEnum.Referer] <- value with _ -> ()
 #else
-            | "Referer" -> req.Referer <- value
+            | "referer" -> req.Referer <- value
 #endif            
-            | "TE" -> req.Headers.[HeaderEnum.Te] <- value
-            | "Trailer" -> req.Headers.[HeaderEnum.Trailer] <- value
-            | "Translate" -> req.Headers.[HeaderEnum.Translate] <- value
-            | "Upgrade" -> req.Headers.[HeaderEnum.Upgrade] <- value
+            | "te" -> req.Headers.[HeaderEnum.Te] <- value
+            | "trailer" -> req.Headers.[HeaderEnum.Trailer] <- value
+            | "translate" -> req.Headers.[HeaderEnum.Translate] <- value
+            | "upgrade" -> req.Headers.[HeaderEnum.Upgrade] <- value
 #if FX_NO_WEBREQUEST_USERAGENT
-            | "User-Agent" -> req.Headers.[HeaderEnum.UserAgent] <- value
+            | "user-agent" -> if not (req?UserAgent <- value) then try req.Headers.[HeaderEnum.UserAgent] <- value with _ -> ()
 #else
-            | "User-Agent" -> req.UserAgent <- value
+            | "user-agent" -> req.UserAgent <- value
 #endif
-            | "Via" -> req.Headers.[HeaderEnum.Via] <- value
-            | "Warning" -> req.Headers.[HeaderEnum.Warning] <- value
-            | name -> req.Headers.[name] <- value))
+            | "via" -> req.Headers.[HeaderEnum.Via] <- value
+            | "warning" -> req.Headers.[HeaderEnum.Warning] <- value
+            | _ -> req.Headers.[header] <- value))
         hasContentType.Value
 
     let getResponse (req:HttpWebRequest) silentHttpErrors =
@@ -436,7 +482,27 @@ module private Helpers =
         else 
             Async.FromBeginEnd(req.BeginGetResponse, req.EndGetResponse)
 
-    let toHttpResponse forceText responseUrl statusCode contentType (_contentEncoding:string)
+    // No inlining to don't cause a depency on ZLib.Portable when a PCL version of FSharp.Data is used in full .NET
+    [<MethodImpl(MethodImplOptions.NoInlining)>]
+    let decompressGZip (memoryStream:MemoryStream) =
+#if FX_NO_WEBREQUEST_AUTOMATICDECOMPRESSION
+        new MemoryStream(Ionic.Zlib.GZipStream.UncompressBuffer(memoryStream.ToArray()))
+#else
+        failwith "Automatic gzip decompression failed"
+        memoryStream
+#endif
+
+    // No inlining to don't cause a depency on ZLib.Portable when a PCL version of FSharp.Data is used in full .NET
+    [<MethodImpl(MethodImplOptions.NoInlining)>]
+    let decompressDeflate (memoryStream:MemoryStream) =
+#if FX_NO_WEBREQUEST_AUTOMATICDECOMPRESSION
+        new MemoryStream(Ionic.Zlib.DeflateStream.UncompressBuffer(memoryStream.ToArray()))
+#else
+        failwith "Automatic deflate decompression failed"
+        memoryStream
+#endif
+
+    let toHttpResponse forceText responseUrl statusCode contentType contentEncoding
                        characterSet responseEncodingOverride cookies headers stream = async {
 
         let isText (mimeType:string) =
@@ -455,15 +521,11 @@ module private Helpers =
         use stream = stream
         let! memoryStream = asyncRead stream
 
-#if FX_NO_WEBREQUEST_AUTOMATICDECOMPRESSION
         let memoryStream = 
-            if _contentEncoding = "gzip" then
-                new MemoryStream(Ionic.Zlib.GZipStream.UncompressBuffer(memoryStream.ToArray()))
-            elif _contentEncoding = "deflate" then
-                new MemoryStream(Ionic.Zlib.DeflateStream.UncompressBuffer(memoryStream.ToArray()))
-            else
-                memoryStream
-#endif
+            // this only applies when automatic decompression is off
+            if contentEncoding = "gzip" then decompressGZip memoryStream
+            elif contentEncoding = "deflate" then decompressDeflate memoryStream
+            else memoryStream
 
         let respBody = 
             if forceText || isText contentType then
@@ -514,8 +576,12 @@ type Http private() =
         // set headers
         let hasContentType = setHeaders headers req
 
+        let automaticDecompression = ref true
+
     #if FX_NO_WEBREQUEST_AUTOMATICDECOMPRESSION
-        req.Headers.[HeaderEnum.AcceptEncoding] <- "gzip,deflate"
+        if not (req?AutomaticDecompression <- 3) then 
+            automaticDecompression := false
+            req.Headers.[HeaderEnum.AcceptEncoding] <- "gzip,deflate"
     #else
         req.AutomaticDecompression <- DecompressionMethods.GZip ||| DecompressionMethods.Deflate
     #endif
@@ -576,13 +642,16 @@ type Http private() =
                 match resp with
                 | :? HttpWebResponse as resp -> 
 #if FX_NO_WEBRESPONSE_CHARACTERSET
-                    int resp.StatusCode, ""
+                    int resp.StatusCode, (defaultArg resp?CharacterSet "")
 #else
                     int resp.StatusCode, resp.CharacterSet
 #endif
                 | _ -> 0, ""
 
-            let contentEncoding = defaultArg (Map.tryFind "Content-Encoding" headers) ""
+            let contentEncoding = 
+                if !automaticDecompression
+                then "" 
+                else defaultArg (Map.tryFind HttpResponseHeaders.ContentEncoding headers) ""
 
             let stream = resp.GetResponseStream()
 
@@ -618,7 +687,19 @@ type Http private() =
     /// The body for POST request can be specified either as text or as a list of parameters
     /// that will be encoded, and the method will automatically be set if not specified
     static member AsyncRequestStream(url, ?query, ?headers, ?httpMethod, ?body, ?cookies, ?cookieContainer, ?silentHttpErrors, ?customizeHttpRequest) =
-        let toHttpResponse responseUrl statusCode _contentType _contentEncoding _characterSet _responseEncodingOverride cookies headers stream = async {
+        let toHttpResponse responseUrl statusCode _contentType contentEncoding _characterSet _responseEncodingOverride cookies headers stream = async {
+            let! stream = async {
+                // this only applies when automatic decompression is off
+                if contentEncoding = "gzip" then 
+                    use stream = stream
+                    let! memoryStream = asyncRead stream
+                    return decompressGZip memoryStream :> Stream
+                elif contentEncoding = "deflate" then 
+                    use stream = stream
+                    let! memoryStream = asyncRead stream
+                    return decompressDeflate memoryStream :> Stream
+                else
+                    return stream }
             return { ResponseStream = stream
                      StatusCode = statusCode
                      ResponseUrl = responseUrl